Move the tree_sitter_like_traverse test to examples dir
